feat: add transitionSpan field to TransitionSnippet struct
- Add uint8 transitionSpan field to TransitionSnippet struct
- Change finalizationDeadline from uint48 to uint40 to fit in bytes32
- Add DeadlineExceedsMax error and validation in _encodeTransitionSnippet
- Update _encodeTransitionSnippet to encode all 3 fields into bytes32
- Update _decodeTransitionSnippet to decode all 3 fields
- Update _getTransitionSnippet return signature to include transitionSpan
- Update all call sites to handle the new return values

+++ packages/protocol/test/layer1/core/inbox/prove/InboxOptimized1TransitionRecord.t.sol
@@ -152,7 +152,7 @@ contract InboxOptimized1TransitionRecord is InboxTestHelper {
         // Verify finalization deadline was set to max via ring buffer
         (uint48 conflictDeadline, bytes26 conflictRecordHash) =
             inbox.getTransitionRecordHash(proposal.id, _getGenesisTransitionHash());
-        assertEq(conflictDeadline, type(uint48).max, "Deadline should be set to max on conflict");
+        assertEq(conflictDeadline, type(uint40).max, "Deadline should be set to max on conflict");
         assertEq(
             conflictRecordHash,
             firstRecordHash,
